Fixed iointr
PyUnicode_AsASCIIString returns a Python Bytes object
The code is now a bit longer. It gives a bit of a message
at each spot. I assume that the code wors as long as
the names given match ASCII definitions.

Perhaps one should consider to force the use of bytes objects
e.g. b'bar'. In this manner one would be sure that the
names in the ioc db match the ones that python sees.